server/requirements.txt
grpcio
grpcio-tools
protobuf
psycopg2-binary
python-dateutil
pydantic
python-json-logger
flask